+++ lib/autoparts/packages/clasp.rb
@@ -2,18 +2,18 @@ module Autoparts
   module Packages
     class Clasp < Package
       name 'clasp'
-      version '3.0.4'
+      version '3.1.1'
       description 'Clasp: an answer set solver for (extended) normal and disjunctive logic programs.'
       category Category::LIBRARIES
 
-      source_url 'http://cznic.dl.sourceforge.net/project/potassco/clasp/3.0.4/clasp-3.0.4-x86_64-linux.tar.gz'
-      source_sha1 '4248a4e77944acc06b2e38a619ebed3a02a4c4e7'
+      source_url 'http://sourceforge.net/projects/potassco/files/clasp/3.1.1/clasp-3.1.1-x86_64-linux.tar.gz'
+      source_sha1 'ded622f8dafb7318f968d75ff34d2d316d581b42'
       source_filetype 'tar.gz'
 
       def install
         FileUtils.mkdir_p(bin_path)
-        Dir.chdir(extracted_archive_path + 'clasp-3.0.4') do
-          execute 'cp', 'clasp-3.0.4-x86_64-linux', bin_path + 'clasp'
+        Dir.chdir(extracted_archive_path + 'clasp-3.1.1') do
+          execute 'cp', 'clasp-3.1.1-x86_64-linux', bin_path + 'clasp'
         end
       end
     end