+++ lib/autoparts/packages/lua.rb
@@ -5,22 +5,22 @@ module Autoparts
   module Packages
     class Lua < Package
       name 'lua'
-      version '5.2.2'
+      version '5.2.3'
       description 'Lua: A powerful, fast, lightweight, embeddable scripting language'
       category Category::PROGRAMMING_LANGUAGES
 
-      source_url 'http://www.lua.org/ftp/lua-5.2.2.tar.gz'
-      source_sha1 '0857e41e5579726a4cb96732e80d7aa47165eaf5'
+      source_url 'http://www.lua.org/ftp/lua-5.2.3.tar.gz'
+      source_sha1 '926b7907bc8d274e063d42804666b40a3f3c124c'
       source_filetype 'tar.gz'
 
       def compile
-        Dir.chdir('lua-5.2.2') do
+        Dir.chdir('lua-5.2.3') do
           execute 'make', 'linux'
         end
       end
 
       def install
-        Dir.chdir('lua-5.2.2') do
+        Dir.chdir('lua-5.2.3') do
           execute 'make', 'install', "INSTALL_TOP=#{prefix_path}", "INSTALL_MAN=#{man_path}/man1"
         end
       end