+++ lib/autoparts/packages/mc.rb
@@ -2,12 +2,12 @@ module Autoparts
   module Packages
     class Mc < Package
       name 'mc'
-      version '4.8.12'
+      version '4.8.13'
       description 'GNU Midnight Commander: a visual file manager'
       category Category::UTILITIES
       
-      source_url 'http://ftp.midnight-commander.org/mc-4.8.12.tar.bz2'
-      source_sha1 '50a95e5e4d272f7cf7821febe238e4e6d4209f34'
+      source_url 'http://ftp.midnight-commander.org/mc-4.8.13.tar.bz2'
+      source_sha1 'ca47a0b824a4675a29eb5ea6862712ef8cf93463'
       source_filetype 'tar.bz2'
 
       depends_on 'glib'